fix compat

+++ migrations/20241230133452_stats_lag_default.up.sql
@@ -1,8 +1,8 @@
 CREATE OR REPLACE VIEW wireguard_peer_stats_view AS
     SELECT
         device_id,
-        greatest(upload - lag(upload, 1, 0) OVER (PARTITION BY device_id ORDER BY collected_at), 0) upload,
-        greatest(download - lag(download, 1, 0) OVER (PARTITION BY device_id ORDER BY collected_at), 0) download,
+        greatest(upload - lag(upload, 1, 0::bigint) OVER (PARTITION BY device_id ORDER BY collected_at), 0) upload,
+        greatest(download - lag(download, 1, 0::bigint) OVER (PARTITION BY device_id ORDER BY collected_at), 0) download,
         latest_handshake - (lag(latest_handshake, 1, latest_handshake) OVER (PARTITION BY device_id ORDER BY collected_at)) latest_handshake_diff,
         latest_handshake,
         collected_at,
